Fixed compilation error in newer compilers.
An explicit cast of hex numbers to char was added in list initializations
of strings. Evidently, the implicit conversion implied in this initializaion
is no longer allowed in list initializations. The code is only relevant to
Chinese TB punctuation, but caused a fatal compilation error in newer
versions of the c++ compiler.